django_server/requirements.txt
asgiref==3.7.2
boto3==1.34.54
botocore==1.34.54
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
mysqlclient==2.2.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.1
s3transfer==0.10.0
sendgrid==6.11.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
urllib3==2.0.7
wheel==0.41.2
langchain==0.1.12
langchain-openai==0.0.8
tiktoken==0.6.0
langchain-core==0.1.31
pandas
neo4j==5.18.0
openai==1.14.1
python-dotenv
googlemaps==4.5.3
stream-chat==4.15.0
selenium
BeautifulSoup4
psycopg2
webdriver_manager
streamlit 
requests 
langchain 
langchain-community 
google-auth 
google-auth-oauthlib 
google-generativeai

